renaming files for consistency sake
