+++ src/models/contributions/index.ts
@@ -1,1 +1,1 @@
-export * from './IVSIcons';
+export * from './vsicons';
