--- src/models/extensions/FileFormat.ts
@@ -1,9 +1,0 @@
-export enum FileFormat {
-  svg,
-  png,
-  jpg,
-  gif,
-  bmp,
-  tiff,
-  ico,
-}
