--- src/models/extensions/SpecialExtensionType.ts
@@ -1,8 +1,0 @@
-export enum SpecialExtensionType {
-  file,
-  folder,
-  folder_open,
-  file_light,
-  folder_light,
-  folder_open_light,
-}
