+++ src/models/extensions/index.ts
@@ -1,9 +1,9 @@
-export * from './FileFormat';
-export * from './IExtension';
-export * from './IExtensionCollection';
-export * from './IFileExtension';
-export * from './IFolderExtension';
-export * from './ISpecialExtension';
-export * from './ISpecialFileCollection';
-export * from './ISpecialFolderCollection';
-export * from './SpecialExtensionType';
+export * from './fileFormat';
+export * from './extension';
+export * from './extensionCollection';
+export * from './fileExtension';
+export * from './folderExtension';
+export * from './specialExtension';
+export * from './specialFileCollection';
+export * from './specialFolderCollection';
+export * from './specialExtensionType';
