+++ src/models/iconGenerator/iconGenerator.ts
