+++ src/models/iconSchema/iconPath.ts
