+++ src/models/iconSchema/index.ts
@@ -1,3 +1,3 @@
-export * from './IIconPath';
-export * from './IIconSchema';
-export * from './ISchemaSettings';
+export * from './iconPath';
+export * from './iconSchema';
+export * from './schemaSettings';
