+++ src/models/settings/index.ts
@@ -1,3 +1,5 @@
-export * from './IExtensionSettings';
-export * from './ISettings';
-export * from './IState';
+export * from './extensionSettings';
+export * from './extensionStatus';
+export * from './settings';
+export * from './settingsManager';
+export * from './state';
