+++ src/settings/index.ts
@@ -1,4 +1,2 @@
 export * from './extensionSettings';
-export * from './ExtensionStatus';
-export * from './ISettingsManager';
-export * from './SettingsManager';
+export * from './settingsManager';
